物料管理系统最新代码提交入库保存

+++ src/main/java/com/zgyw/materiel/service/impl/TGroupServiceImpl.java
@@ -1,18 +1,17 @@
 package com.zgyw.materiel.service.impl;
 
 
-import com.zgyw.materiel.bean.Classify;
-import com.zgyw.materiel.bean.TGroup;
+import com.zgyw.materiel.bean.*;
 import com.zgyw.materiel.enums.ResultEnum;
 import com.zgyw.materiel.exception.MTException;
-import com.zgyw.materiel.repository.ClassifyRepository;
-import com.zgyw.materiel.repository.TGroupRepository;
+import com.zgyw.materiel.repository.*;
 import com.zgyw.materiel.service.TGroupService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class TGroupServiceImpl implements TGroupService {
@@ -20,6 +19,12 @@ public class TGroupServiceImpl implements TGroupService {
     private TGroupRepository repository;
     @Autowired
     private ClassifyRepository classifyRepository;
+    @Autowired
+    private MaterielRecordsRepository recordsRepository;
+    @Autowired
+    private MaterielLevelRepository levelRepository;
+    @Autowired
+    private OrderRecordsRepository orderRepository;
 
     @Override
     @Transactional
@@ -55,6 +60,15 @@ public class TGroupServiceImpl implements TGroupService {
     @Transactional
     public void delete(Integer id) {
         List<Classify> list = classifyRepository.findByGroupId(id);
+        List<Integer> classifyIds = list.stream().map(e -> e.getId()).collect(Collectors.toList());
+        List<String> names = list.stream().map(e -> e.getName()).collect(Collectors.toList());
+        List<MaterielLevel> materielLevels = levelRepository.findByClassifyIdIn(classifyIds);
+        List<OrderRecords> orderRecords = orderRepository.findByStatus(0);
+        List<Integer> orderIds = orderRecords.stream().map(e -> e.getId()).collect(Collectors.toList());
+        List<MaterielRecords> materielRecords = recordsRepository.findByOrderIdInAndNameIn(orderIds, names);
+        if (materielLevels.size() > 0 || materielRecords.size() > 0) {
+            throw new MTException(ResultEnum.CLASSIFY_EXIST);
+        }
         classifyRepository.deleteInBatch(list);
         repository.deleteById(id);
     }
